requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Streamlit UI
streamlit==1.32.0

# Video Generation
edge-tts==6.1.18
faster-whisper==1.0.3
Pillow==9.5.0
numpy==1.26.0
pydub==0.25.1
scipy==1.11.4
moviepy==1.0.3

# HTTP Client (for external mode)
requests==2.31.0
httpx==0.26.0